+++ docs/CHANGELOG.md
@@ -49,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove several cvss fields from Affect (OSIDB-2749)
 - Remove "type" field from FlawComment (OSIDB-2745)
 - Remove FlawMeta (OSIDB-2744)
+- Remove "is_major_incident" field from Flaw (OSIDB-2741)
 
 ## [3.7.3] - 2024-05-28
 ### Fixed